debug level software
most hardware specific items are in mux57.c file, as display and keyboard support is the bulk of the embedded hardware task. license file updated to include Jeff Parsons' original MIT as well as MIT for my own files. Astyle tool was used to format my code files in a consistent fashion.